requirements.txt
Flask==3.0.0
pymysql==1.1.0
python-dotenv==1.0.1
cryptography==42.0.5
